@@ -9,24 +9,20 @@
 	"license": "MIT",
 	"type": "project",
 	"require": {
-		"php": ">=7.2",
-		"doctrine/dbal": "^2.8",
-		"flipbox/lumen-generator": "^7.0",
-		"laravel/lumen-framework": "^7.0",
-		"vlucas/phpdotenv": "^4.0"
+		"php": ">=7.3",
+		"flipbox/lumen-generator": "^8.0",
+		"laravel/lumen-framework": "^8.0"
 	},
 	"require-dev": {
 		"fzaninotto/faker": "~1.4",
 		"mockery/mockery": "~1.0",
-		"phpunit/phpunit": "~7.0"
+		"phpunit/phpunit": "^9.0"
 	},
 	"autoload": {
-		"classmap": [
-			"database/seeds",
-			"database/factories"
-		],
 		"psr-4": {
-			"App\\": "app/"
+			"App\\": "app/",
+			"Database\\Factories\\": "database/factories/",
+			"Database\\Seeders\\": "database/seeders/"
 		}
 	},
 	"autoload-dev": {
