@@ -1,10 +1,12 @@
 <?php
 
+namespace Database\Seeders;
+
 use Illuminate\Database\Seeder;
 use App\User;
 use App\Title;
 use App\Chapter;
-use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\DB;
 
 class DatabaseSeeder extends Seeder
 {
@@ -22,8 +24,8 @@ class DatabaseSeeder extends Seeder
         Title::truncate();
         Chapter::truncate();
 
-        factory(User::class, 10)->create()->each(function ($user) {
-            factory(Title::class, 10)->create([
+        User::factory()->count(10)->create()->each(function ($user) {
+            Title::factory()->count(10)->create([
                 'user_id' => $user->id,
             ])->each(function ($title) {
                 $chapters = [];
